[WICKET-6497] wicket-event-jquery.js and wicket-ajax-jquery.js are merged, references are updated
