+++ wicket-core/src/main/java/org/apache/wicket/settings/JavaScriptLibrarySettings.java
@@ -18,9 +18,7 @@ package org.apache.wicket.settings;
 
 import org.apache.wicket.ajax.WicketAjaxDebugJQueryResourceReference;
 import org.apache.wicket.ajax.WicketAjaxJQueryResourceReference;
-import org.apache.wicket.ajax.WicketEventJQueryResourceReference;
 import org.apache.wicket.request.resource.ResourceReference;
-import org.apache.wicket.resource.DynamicJQueryResourceReference;
 import org.apache.wicket.resource.JQueryResourceReference;
 import org.apache.wicket.util.lang.Args;
 
@@ -38,9 +36,7 @@ import org.apache.wicket.util.lang.Args;
  */
 public class JavaScriptLibrarySettings
 {
-	private ResourceReference jQueryReference = DynamicJQueryResourceReference.getV2();
-
-	private ResourceReference wicketEventReference = WicketEventJQueryResourceReference.get();
+	private ResourceReference jQueryReference = JQueryResourceReference.getV2();
 
 	private ResourceReference wicketAjaxReference = WicketAjaxJQueryResourceReference.get();
 
@@ -64,25 +60,6 @@ public class JavaScriptLibrarySettings
 	public JavaScriptLibrarySettings setJQueryReference(ResourceReference jQueryReference)
 	{
 		this.jQueryReference = Args.notNull(jQueryReference, "jQueryReference");
-		return this;
-	}
-
-	/**
-	 * @return the reference to the implementation of wicket-event.js
-	 */
-	public ResourceReference getWicketEventReference()
-	{
-		return wicketEventReference;
-	}
-
-	/**
-	 * @param wicketEventReference
-	 *            a reference to the implementation of wicket-event.js
-	 * @return {@code this} object for chaining
-	 */
-	public JavaScriptLibrarySettings setWicketEventReference(ResourceReference wicketEventReference)
-	{
-		this.wicketEventReference = Args.notNull(wicketEventReference, "wicketEventReference");
 		return this;
 	}
 
@@ -126,5 +103,4 @@ public class JavaScriptLibrarySettings
 			"wicketAjaxDebugReference");
 		return this;
 	}
-
 }
