+++ wicket-core/src/test/java/org/apache/wicket/ajax/AjaxHeaderContributionPage_ajax_expected.html
@@ -8,5 +8,5 @@ test
 <link href="../../test"/>
 <script type="text/javascript" src="../../javascripturl"></script>
 <script type="text/javascript" src="../resource/org.apache.wicket.resource.JQueryResourceReference/jquery/jquery-2.2.4.js"></script>
-<script type="text/javascript" src="../resource/org.apache.wicket.ajax.AbstractDefaultAjaxBehavior/res/js/wicket-event-jquery.js"></script>
+<script type="text/javascript" src="../resource/org.apache.wicket.ajax.AbstractDefaultAjaxBehavior/res/js/wicket-ajax-jquery.js"></script>
 </head>]]></header-contribution><priority-evaluate><![CDATA[(function(){prepend();})();]]></priority-evaluate><evaluate><![CDATA[(function(){domReady();})();(function(){append();})();(function(){onLoad();})();]]></evaluate></ajax-response>
