+++ wicket-core/src/test/js/all.html
@@ -9,7 +9,6 @@
         document.write("<scr"+"ipt src='../../main/java/org/apache/wicket/resource/jquery/jquery-"+location.search.substring(1)+".js'></scr"+"ipt>");
     </script>
     <!--<script src="https://code.jquery.com/jquery-migrate-3.0.0.js"></script>-->
-    <script src="../../main/java/org/apache/wicket/ajax/res/js/wicket-event-jquery.js"></script>
 	<script src="../../main/java/org/apache/wicket/ajax/res/js/wicket-ajax-jquery.js"></script>
 	<script src="../../main/java/org/apache/wicket/ajax/res/js/wicket-ajax-jquery-debug.js"></script>
 	<script type="text/javascript" src="qunit/qunit.js"></script>
