+++ wicket-core/src/test/js/amd.js
@@ -35,8 +35,8 @@ define(
 					qunit.expect(15);
 
 					// test for one member for each Wicket.** module
-					qunit.ok($.isFunction(w.Browser.isChrome), "Wicket Browser (from wicket-event-jquery.js) is imported");
-					qunit.ok($.isFunction(w.Event.fix), "Wicket Event (from wicket-event-jquery.js) is imported");
+					qunit.ok($.isFunction(w.Browser.isChrome), "Wicket Browser (from wicket-ajax-jquery.js) is imported");
+					qunit.ok($.isFunction(w.Event.fix), "Wicket Event (from wicket-ajax-jquery.js) is imported");
 					qunit.ok($.isFunction(w.Ajax.ajax), "Wicket Ajax (from wicket-ajax-jquery.js) is imported");
 					qunit.ok($.isFunction(w.DOM.get), "Wicket DOM (from wicket-ajax-jquery.js) is imported");
 					qunit.ok($.isFunction(w.Form.serialize), "Wicket Form (from wicket-ajax-jquery.js) is imported");
